# internal/docs/swagger/openapi.yaml
openapi: 3.0.0
info:
  title: Sarbon API
  description: |
    Закрытый backend — доступ только для доверенных клиентов (frontend или mobile).
    На каждый запрос обязательны заголовки (токен только в header X-Client-Token):
    - X-Client-Type (frontend | mobile)
    - X-Platform (web | ios | android)
    - X-Client-Token — токен приложения (обязателен для всех API)
    - Accept-Language (ru | en | uz | tr | zh)
    Authorization: Bearer &lt;JWT&gt; — только для API, где в описании указано «требуется JWT».
  version: 1.0.0
  x-response-format: |
    Все ответы API (JSON) имеют единый формат: status (число — HTTP-код), message (строка — "success" или текст ошибки), data (тело ответа или null при ошибке).
servers:
  - url: /api/v1
    description: API v1
...
paths:
...
  /drivers/profile:
    get:
      summary: Получить свой профиль
      description: Возвращает все данные текущего водителя из таблицы (по JWT). Требуется Client token + Bearer.
      operationId: getDriverProfile
      tags: [Drivers Profile]
      security:
        - ClientToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Профиль водителя (все поля).
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: integer, example: 200 }
                  message: { type: string }
                  data:
                    $ref: '#/components/schemas/Driver'
        '401':
          description: unauthorized
        '404':
          description: Водитель не найден
        '500':
          description: Внутренняя ошибка
    put:
      summary: Обновить профиль (полностью)
      description: Обновляет все переданные поля. Номер телефона изменить нельзя. JSON или multipart (файлы car_photo, adr_document; remove_car_photo=1, remove_adr_document=1). Язык и платформа из заголовков.
      operationId: updateDriverProfile
      tags: [Drivers Profile]
      security:
        - ClientToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDriverProfileRequest'
          multipart/form-data:
            schema:
              type: object
              properties:
                first_name: { type: string }
                last_name: { type: string }
                passport_series: { type: string }
                passport_number: { type: string }
                company_id: { type: string, format: uuid }
                freelance_dispatcher_id: { type: string, format: uuid }
                car_photo: { type: string, format: binary }
                adr_document: { type: string, format: binary }
                remove_car_photo: { type: string }
                remove_adr_document: { type: string }
      responses:
        '200':
          description: Обновлено.
        '401':
          description: unauthorized
        '404':
          description: Водитель не найден
        '500':
          description: Внутренняя ошибка
    patch:
      summary: Обновить профиль (частично)
      description: Обновляет только переданные поля. Номер телефона изменить нельзя.
      operationId: patchDriverProfile
      tags: [Drivers Profile]
      security:
        - ClientToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateDriverProfileRequest'
      responses:
        '200':
          description: Обновлено.
        '401':
          description: unauthorized
        '404':
          description: Водитель не найден
        '500':
          description: Внутренняя ошибка
    delete:
      summary: Удалить свой профиль
      description: Переносит запись в deleted_drivers (архив), затем удаляет из drivers. Требуется Bearer.
      operationId: deleteDriverProfile
      tags: [Drivers Profile]
      security:
        - ClientToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Профиль удалён.
        '401':
          description: unauthorized
        '404':
          description: Водитель не найден
        '500':
          description: Внутренняя ошибка
  /drivers/profile/last-activate:
    get:
      summary: Получить последнюю активацию
      description: Возвращает время и координаты последней активации (last_activated_at, last_activated_latitude, last_activated_longitude). Требуется Bearer.
      operationId: getDriverProfileLastActivate
      tags: [Drivers Profile]
      security:
        - ClientToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Время и координаты последней активации (в data).
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: integer, example: 200 }
                  message: { type: string }
                  data:
                    type: object
                    properties:
                      last_activated_at: { type: string, format: date-time }
                      last_activated_latitude: { type: number }
                      last_activated_longitude: { type: number }
        '401':
          description: unauthorized
        '404':
          description: Водитель не найден
        '500':
          description: Внутренняя ошибка
    patch:
      summary: Обновить последнюю активацию
      description: Время и координаты последней активации на карте. Требуется Bearer.
      operationId: patchDriverProfileLastActivate
      tags: [Drivers Profile]
      security:
        - ClientToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                activated_at: { type: string, format: date-time }
                latitude: { type: number }
                longitude: { type: number }
      responses:
        '200':
          description: Обновлено.
        '401':
          description: unauthorized
        '500':
          description: Внутренняя ошибка
  /drivers/profile/files/car-photo:
    get:
      summary: Скачать фото машины (свой профиль)
      description: Возвращает файл по ссылке из GET /drivers/profile (car_photo_url). Bearer обязателен.
      operationId: getDriverProfileCarPhoto
      tags: [Drivers Profile]
      security:
        - ClientToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Файл (изображение)
        '401':
          description: unauthorized
        '404':
          description: Файл не найден
        '500':
          description: Внутренняя ошибка
  /drivers/profile/files/adr-document:
    get:
      summary: Скачать ADR-документ (свой профиль)
      description: Возвращает файл по ссылке из GET /drivers/profile (adr_document_url). Bearer обязателен.
      operationId: getDriverProfileAdrDocument
      tags: [Drivers Profile]
      security:
        - ClientToken: []
        - BearerAuth: []
      parameters:
        - $ref: '#/components/parameters/XClientType'
        - $ref: '#/components/parameters/XPlatform'
        - $ref: '#/components/parameters/AcceptLanguage'
      responses:
        '200':
          description: Файл (PDF/изображение)
        '401':
          description: unauthorized
        '404':
          description: Файл не найден
        '500':
          description: Внутренняя ошибка
components:
  schemas:
...
    Driver:
      type: object
      description: |
        Модель водителя (ответ API).
        Поля created_at, updated_at, deleted_at задаются только бэкендом — фронт не передаёт и не изменяет их.
        created_at — при создании; updated_at — при PUT/PATCH; deleted_at — при мягком удалении (DELETE).
      properties:
        id:
          type: string
          format: uuid
        first_name:
          type: string
        last_name:
          type: string
        phone_number:
          type: string
        passport_series:
          type: string
        passport_number:
          type: string
        company_id:
          type: string
          format: uuid
          nullable: true
        freelance_dispatcher_id:
          type: string
          format: uuid
          nullable: true
          description: ID фриланс-диспетчера (как company_id)
        car_photo_url:
          type: string
          nullable: true
        adr_document_url:
          type: string
          nullable: true
        rating:
          type: number
          format: float
          minimum: 0
          maximum: 5
        work_status:
          type: string
          enum: [free, busy]
        account_status:
          type: string
          enum: [pending, approved, blocked]
        language:
          type: string
          enum: [ru, uz, en, tr, zh]
        platform:
          type: string
          description: Из заголовка X-Platform (web, ios, android)
        dispatcher_type:
          type: string
          description: Тип диспетчера
        last_activated_at:
          type: string
          format: date-time
          nullable: true
          description: Время последней активации (PATCH /drivers/profile/last-activate)
        last_activated_latitude:
          type: number
          format: double
          nullable: true
          description: Широта последней активации на карте
        last_activated_longitude:
          type: number
          format: double
          nullable: true
          description: Долгота последней активации на карте
        created_at:
          type: string
          format: date-time
          description: Только бэкенд (при создании)
        updated_at:
          type: string
          format: date-time
          description: Только бэкенд (при PUT/PATCH)
        deleted_at:
          type: string
          format: date-time
          nullable: true
          description: Только бэкенд (при мягком удалении)
...
    UpdateDriverProfileRequest:
      type: object
      description: Поля профиля водителя (номер телефона изменить нельзя). Язык и платформа из заголовков.
      properties:
        first_name: { type: string }
        last_name: { type: string }
        passport_series: { type: string }
        passport_number: { type: string }
        company_id: { type: string, format: uuid, nullable: true }
        freelance_dispatcher_id: { type: string, format: uuid, nullable: true }
        rating: { type: number, minimum: 0, maximum: 5 }
        work_status: { type: string, enum: [free, busy] }
        account_status: { type: string, enum: [pending, approved, blocked] }
        language: { type: string, enum: [ru, uz, en, tr, zh] }
        platform: { type: string }
        dispatcher_type: { type: string }
...
  securitySchemes:
    ClientToken:
      type: apiKey
      in: header
      name: X-Client-Token
      description: Токен приложения — обязателен для каждого запроса. Укажи FRONTEND_CLIENT_TOKEN или MOBILE_CLIENT_TOKEN.
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT пользователя — только для API с пометкой «требуется JWT».
  parameters:
    XClientType:
      name: X-Client-Type
      in: header
      required: true
      schema:
        type: string
        enum: [frontend, mobile]
      description: Тип клиента
    XPlatform:
      name: X-Platform
      in: header
      required: true
      schema:
        type: string
        enum: [web, ios, android]
      description: Платформа
    AcceptLanguage:
      name: Accept-Language
      in: header
      required: true
      schema:
        type: string
        enum: [ru, en, uz, tr, zh]
      description: Язык (только из заголовка)